Trial a more modular pattern directory structure for pages and page-sets.
